requirements.txt
Django>=2.2.5
django-cors-headers==3.7.0
django-mathfilters==1.0.0
django-picklefield==3.0.1
django-widget-tweaks==1.4.8
Pillow>=7.1.2
razorpay==1.2.0